requirements.txt
numpy==1.24.3
pandas==2.0.3
flask==2.0.1
werkzeug==2.0.1
python-dotenv==1.0.0
requests==2.28.2
flask-login==0.6.0
flask-wtf==1.1.0
anthropic==0.3.0
httpx==0.24.1
pydantic==1.10.8